# Combine `CharBuffer`/`StringBuffer` classes, and use in more places