… core/string/string_buffer.h