- Made `StringBuffer` character type templated, to support both `char` and `char32_t` buffers
- Trimmed down implementation to focus on only appending one character at a time (other methods were unused)
- Replaced `CharBuffer` class from `file_access.cpp`, and renamed `StringBuffer` to `CharBuffer`
- Updated JSON parsing and String uri encoding/decoding to use `CharBuffer`

This makes `JSON.parse_string` around 50% faster, and uri encode/decode methods around 2x faster, compared with gdscript below. `ConfigFile.load` and `FileAccess.get_line` are also included below just to check that using `CharBuffer<>` isn't slower than the previous implementations.

```gdscript
func _ready() -> void:
	run_test("json_parse")
	run_test("uri_encode_decode")
	run_test("load_config")
	run_test("read_lines")
	
func run_test(p_name : String):
	var start := Time.get_ticks_msec()
	call(p_name)
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [p_name, end - start])

func json_parse():
	var d := {
		"key1": 1,
		"key2": "string value",
		"key3": 1.2345,
		"key4": true
	}
	var str := JSON.stringify(d)
	for i in 100000:
		var d2 = JSON.parse_string(str)

func uri_encode_decode():
	var query_params := "is%3Apr+is%3Aopen+label%3Atopic%3Arendering"
	for i in 200000:
		query_params.uri_decode().uri_encode()
		
func load_config():
	for i in 5000:
		var config := ConfigFile.new()
		config.load("res://icon.svg.import")

func read_lines():
	for i in 5000:
		var file_access := FileAccess.open("res://icon.svg.import", FileAccess.READ)
		while !file_access.eof_reached():
			file_access.get_line()
```

old:
```
json_parse: 254ms
uri_encode_decode: 417ms
load_config: 456ms
read_lines: 389ms
```

new:
```
json_parse: 156ms
uri_encode_decode: 182ms
load_config: 436ms
read_lines: 388ms
```

This change started out with looking into how to combine `StringBuffer`/`StringBuilder`, and I came across https://github.com/godotengine/godot/pull/77158 which was looking into a similar change.  After trying out some ideas and seeing some of the benchmarks there, it felt like while the two classes are similar, they have distinct goals - `StringBuffer` more focused on reducing/eliminating allocations for small strings, and `StringBuilder` more focused on handling longer strings.  Rather than trying to come up with an approach that works well for both cases, this PR attempts to trim down `StringBuffer` (now `CharBuffer`) to focus on the case of building small strings out of individual characters, and reuse the class in a few more places that can benefit from it.  Going this route would also hopefully make optimizing `StringBuilder` easier in the future, without having to worry about trying to keep it fast in the cases where `CharBuffer` would be a better fit

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->