@aaronp64 2025-05-19
Do you mean use only `String`/`CharStringT`, or keep the fixed array buffer, and only replace the `LocalVector` part?  I think keeping the array is useful, as there are some cases where we're parsing numbers, and no allocations are done at all.

My initial feeling about replacing the `LocalVector` is the extra complexity would outweigh the performance gains, though I can try it out and compare.  The cases where the `LocalVector` (or `String`/`CharStringT`) actually gets used should be rare - if it's happening enough to show up as a bottleneck, the `FIXED_BUFFER_SIZE` should probably be changed, or just not use `CharBuffer`.  Though I guess when string types are being read in, we can't really know the length for sure.  The choice of `LocalVector` was based on just using the fixed buffer in the common case, and being simple and "good enough" for when the string is longer than expected.

Regarding the `finalize` method - the current `CharBuffer` implementation allows getting the current state of the string, then continuing to append to it, though this is currently unused.  I was going to say `finalize` would make that case slower, but I think avoiding the copy when getting the first string state would cancel that out.  It would be another step to handle that though if we want to support that case.