@aaronp64 2025-05-19
I did some benchmarking by changing the `json_parse` dictionary to the below, to test what happens with longer strings:

```gdscript
	var d := {
		"key1": 1,
		"key2": "long string".repeat(1000),
		"key3": 1.2345,
		"key4": true
	}
```

On `master`, this took 13061ms.  With the current PR code, it took 2258ms.  Some of the time taken with the current PR code is due to `String::append_utf32` validating each character - adding the below and using `as_string()` instead of `get_terminated_buffer()` reduced the time to 1810ms:

```c++
	using StringType = std::conditional_t<std::is_same_v<T, char32_t>, String, CharStringT<T>>;
	...
	const StringType as_string() {
		T *current_buffer = _get_current_buffer();
		current_buffer[_length] = '\0';

		StringType result;
		result.resize(_length + 1);
		T *result_ptr = result.ptrw();
		memcpy(result_ptr, current_buffer, (_length + 1) * sizeof(T));
		return result;
	}
```

This raises the question - do we want `CharBuffer` to validate the string, have it as an option, or no validation?  The original `StringBuffer` didn't validate, while the json code updated by this PR did validate.  The current PR code validates through `String(str.get_terminated_buffer())`.

Changing to use `String/CharStringT` inside `CharBuffer` reduces the time further to 1561ms.  In addition to the changes you mentioned in your earlier comment, I also added `CowData::set_size` to manually set the size, without risking `resize` zeroing out/overwriting the characters.  I think the current implementation doesn't actually require that, between `p_ensure_zero` and `std::is_trivially_constructible_v<T>` checks, but I'm not sure how much we want to rely on `CowData` implementation details from `CharBuffer`.  There's also an expectation that `CowData::resize` will always choose capacity based on some growth factor, and not allocate an exact amount.  If we `resize` with a growth factor from `CharBuffer`, we would have to `resize` down to the final length at the end, which could cause the same copy anyway.  At least, this is based on what I currently have for `_reserve` when trying to implement this:

```c++
	void _reserve(size_t p_size) {
		if (p_size <= _capacity) {
			return;
		}

		bool copy = !_dynamic_buffer && _length > 0;
		if (!copy) {
			_string.set_size(_length + 1);
		}
		_string.resize(p_size);
		_capacity = _string.capacity();
		_dynamic_buffer = _string.ptrw();
		if (copy) {
			memcpy(_dynamic_buffer, _fixed_buffer, _length * sizeof(T));
		}
	}
```

Summary of times/string validation details:
|| master | current PR | PR with `as_string` | PR with `String` inside `CharBuffer` |
| --- | --- | --- | --- | --- |
| json_parse time | 13061ms | 2258ms | 1810ms | 1561ms |
| string validated? | no when `StringBuffer` was used, yes for other `char32_t`/`String` cases | yes | no | no |

The performance difference was larger than I was expecting.  I think at this point the big questions are what we want to do with string validation, and how much we want to expose/rely on cowdata implementation.