@Ivorforce 2025-05-19
Thanks for all those extensive tests! I think they show the potential for a high performance API like this :)

> I think at this point the big questions are what we want to do with string validation

Practically, we probably want a similar API like `String`, where you can append either fast (i.e. unchecked), or with some format in mind (e.g. `append_utf8`).
I'm not sure what the best approach is to avoid duplicating code. Perhaps we can start by doing no validation whatsoever, and making certain that users of `CharBuffer` only append valid string parts.

> how much we want to expose/rely on cowdata implementation.

What do you mean by this? I'd hope the coupling would be relatively low, except for the `const_cast` on `ptr` and the `capacity()` call to ensure no size differences.

Side ramble: Actually, thinking about this now, this whole `CharBuffer` idea ties in pretty well with another concern I had recently, where you may want occasional high-performance single reference access to a `CowData` type temporarily. For example, https://github.com/godotengine/godot/pull/105264 recently used a `LocalVector` as a temporary buffer, while switching to `Vector` later. While the right decision at this time, I still think it's kind of kind of silly it requires a finalizing copy.
Anyway, enough about that. It probably makes sense to move forwards with this PR for now, and keep the above mentioned ideas in mind for possible future work.